To support this log scraping is done every hour, but if logs not enabled this has no impact.

+++ unbound_stats.sh
@@ -12,14 +12,21 @@
 ## v1.1.0 - March 3 2020 - Added graphs for histogram and answers, fixed install to not create duplicate tabs
 ## v1.1.1 - March 8 2020 - Added new install of JackYaz shared graphing files (previously needed to have one of JackYaz's other plugins installed)
 ## v1.1.2 - March 9 2020 - Cleanup .db and .md5 files on uninstall, move startup to post-mount, fixed directory check
-## v1.1.3 - April 12 2020 - Fix error message shown about missing md5 file during clean install
-## v1.1.4 - April 13 2020 - During install, do not Generate stats if unbound is not running, move db off jffs to USB Key
-readonly SCRIPT_VERSION="v1.1.4"
+## v1.2.0 - March 23 2020 - Add output for top ad blocked graph top 10 and top domains - moved stats DB to USB
+## v1.2.1 - March 26 2020 - Added daily replies table
+## v1.2.2 - April 5 2020 - Added tracking of client ip
+## v1.2.3 - April 10 2020 - Fixed issue with "" domain name in SQL, breaking JS
+## v1.2.4 - April 12 2020 - Removed error message on clean install for missing md5 file
+## v1.2.5 - April 13 2020 - During install, do not Generate stats if unbound is not running
+## v1.3.0 - April 16 2020 - Show stats for DNS Firewall
+readonly SCRIPT_VERSION="v1.3.0"
 
 #define www script names
 readonly SCRIPT_WEBPAGE_DIR="$(readlink /www/user)"
 readonly SCRIPT_NAME="Unbound_Stats.sh"
+readonly LOGSCRIPT_NAME="Unbound_Log.sh"
 readonly SCRIPT_NAME_LOWER="unbound_stats.sh"
+readonly LOGSCRIPT_NAME_LOWER="unbound_log.sh"
 readonly SCRIPT_WEB_DIR="$SCRIPT_WEBPAGE_DIR/$SCRIPT_NAME_LOWER"
 
 readonly SCRIPT_DIR="/jffs/addons/unbound"
@@ -39,13 +46,20 @@ statsTitleFile="$SCRIPT_WEB_DIR/unboundstatstitle.txt"
 statsFileJS="$SCRIPT_WEB_DIR/unboundstats.js"
 statsTitleFileJS="$SCRIPT_WEB_DIR/unboundstatstitle.js"
 statsCHPFileJS="$SCRIPT_WEB_DIR/unboundchpstats.js"
+statsRPZFileJS="$SCRIPT_WEB_DIR/unboundrpzstats.js"
+statsRPZHitsFileJS="$SCRIPT_WEB_DIR/unboundrpzhits.js"
 statsHistogramFileJS="$SCRIPT_WEB_DIR/unboundhistogramstats.js"
 statsAnswersFileJS="$SCRIPT_WEB_DIR/unboundanswersstats.js"
+statsTopBlockedFileJS="$SCRIPT_WEB_DIR/unboundtopblockedstats.js"
+statsTopRepliesFileJS="$SCRIPT_WEB_DIR/unboundtoprepliesstats.js"
+statsDailyRepliesFileJS="$SCRIPT_WEB_DIR/unbounddailyreplies.js"
+dailyRepliesCSVFile="$SCRIPT_WEB_DIR/unboundrepliestoday.csv"
 adblockStatsFile="/opt/var/lib/unbound/adblock/stats.txt"
 
 #DB file to hold data for uptime graph
 dbOldStats="$SCRIPT_DIR/unboundstats.db"
 dbStats="/opt/var/lib/unbound/unbound_stats.db"
+dbLogs="/opt/var/lib/unbound/unbound_log.db"
 
 #save md5 of last installed www ASP file so you can find it again later (in case of www ASP update)
 installedMD5File="$SCRIPT_DIR/www-installed.md5"
@@ -72,7 +86,11 @@ WriteData_ToJS(){
 	contents="$3"'.unshift( '
 	while IFS='' read -r line || [ -n "$line" ]; do
 		if echo "$line" | grep -q "NaN"; then continue; fi
-		datapoint="{ x: moment.unix(""$(echo "$line" | awk 'BEGIN{FS=","}{ print $1 }' | awk '{$1=$1};1')""), y: ""$(echo "$line" | awk 'BEGIN{FS=","}{ print $2 }' | awk '{$1=$1};1')"" }"
+		if [ "$4" == "date-day" ]; then
+			datapoint="{ x: moment(\"""$(echo "$line" | awk 'BEGIN{FS=","}{ print $1 }' | awk '{$1=$1};1')""\", \"YYYY-MM-DD\"), y: ""$(echo "$line" | awk 'BEGIN{FS=","}{ print $2 }' | awk '{$1=$1};1')"" }"
+		else	
+			datapoint="{ x: moment.unix(""$(echo "$line" | awk 'BEGIN{FS=","}{ print $1 }' | awk '{$1=$1};1')""), y: ""$(echo "$line" | awk 'BEGIN{FS=","}{ print $2 }' | awk '{$1=$1};1')"" }"
+		fi
 		contents="$contents""$datapoint"","
 	done < "$1"
 	contents=$(echo "$contents" | sed 's/.$//')
@@ -121,6 +139,82 @@ WriteUnboundLabels_ToJS(){
 		echo "${outputvar}.unshift($outputlist);"
 		echo; } >> "$outputfile"
 }
+
+#$1sql table $2 label column $3 count column $4 limit count $5 csv file $6 sql file $7 where clasue if needed
+WriteUnboundSqlLog_ToFile(){
+	{
+		echo ".mode csv"
+		echo ".output $5"
+	} > "$6"
+	echo "SELECT $2, SUM($3) FROM $1 $7 GROUP BY $2 ORDER BY SUM($3) DESC LIMIT $4;" >> "$6"
+}
+
+#$1 csv file $2 js file $3 varLabel $4 varData
+WriteUnboundCSV_ToJS() {
+	labels="$3"'.unshift( '
+	values="$4"'.unshift( '
+	while IFS='' read -r line || [ -n "$line" ]; do
+		if echo "$line" | grep -q "NaN"; then continue; fi
+		labels="$labels""$(echo "$line" | awk 'BEGIN{FS=","}{ print "\x27" $1 "\x27" }' | awk '{$1=$1};1')"","
+		values="$values""$(echo "$line" | awk 'BEGIN{FS=","}{ print $2 }' | awk '{$1=$1};1')"","
+	done < "$1"
+	labels=$(echo "$labels" | sed 's/.$//')
+	labels="$labels"");"
+	values=$(echo "$values" | sed 's/.$//')
+	values="$values"");"
+
+	{
+	echo "var $3;"
+	echo "$3 = [];"; } >> "$2"
+	printf "%s\\r\\n\\r\\n" "$labels" >> "$2"
+	{
+	echo "var $4;"
+	echo "$4 = [];"; } >> "$2"
+	printf "%s\\r\\n\\r\\n" "$values" >> "$2"
+}
+
+#$1 csv file $2 js file $3 varLabel $4 varData
+WriteUnboundCSV_ToJS_2Labels() {
+	labels="$3"'.unshift( '
+	values="$4"'.unshift( '
+	while IFS='' read -r line || [ -n "$line" ]; do
+		if echo "$line" | grep -q "NaN"; then continue; fi
+		labels="$labels""$(echo "$line" | awk 'BEGIN{FS=","}{ print "\x27" $1 " (" $2 ")\x27" }' | awk '{$1=$1};1')"","
+		values="$values""$(echo "$line" | awk 'BEGIN{FS=","}{ print $3 }' | awk '{$1=$1};1')"","
+	done < "$1"
+	labels=$(echo "$labels" | sed 's/.$//')
+	labels="$labels"");"
+	values=$(echo "$values" | sed 's/.$//')
+	values="$values"");"
+
+	{
+	echo "var $3;"
+	echo "$3 = [];"; } >> "$2"
+	printf "%s\\r\\n\\r\\n" "$labels" >> "$2"
+	{
+	echo "var $4;"
+	echo "$4 = [];"; } >> "$2"
+	printf "%s\\r\\n\\r\\n" "$values" >> "$2"
+}
+
+#$1 csv file $2 JS file $3 JS func name $4 html tag
+WriteUnboundCSV_ToJS_Table() {
+	#clean up any null (or "") strings with null string
+	sed -i 's/""/null/g' "$1"
+
+	[ -f $2 ] && rm -f "$2"
+	echo "function $3(){" >> "$2"
+	html='document.getElementById("'"$4"'").outerHTML="'
+	numLines="$(wc -l < $1)"
+	if [ "$numLines" -lt 1 ]; then
+		html="$html""<tr><td colspan="4" class="nodata">No data to display</td></tr>"
+	else
+		html="$html""$(cat "$1" | awk 'BEGIN{FS=","}{ print "<tr><td>" $1 "</td><td>" $2 "</td><td>"$3 "</td><td>" $4 "</td></tr> \\" }' | awk '{$1=$1};1')"
+	fi
+	html=${html%?}
+	html="$html"'"'
+	printf "%s}" "$html" >> "$2"
+} 
 
 #$1 fieldname $2 tablename $3 frequency (hours) $4 length (days) $5 outputfile $6 sqlfile
 WriteSql_ToFile(){
@@ -173,7 +267,11 @@ Generate_UnboundStats () {
 	fi
 	
 	#calc % served by cache
-	UNB_CHP="$(awk 'BEGIN {printf "%0.2f", '$UNB_NUM_CH'*100/'$UNB_NUM_Q'}' )"
+	if [ ! -z "$UNB_NUM_Q" ] && [ $UNB_NUM_Q -ne 0 ]; then
+		UNB_CHP="$(awk 'BEGIN {printf "%0.2f", '$UNB_NUM_CH'*100/'$UNB_NUM_Q'}' )"
+	else
+		UNB_CHP=0
+	fi
 	echo "Calculated Cache Hit Percentage: $UNB_CHP"
 	printf "$(awk 'BEGIN {printf "\\n\\n Cache hit success percent: %s", '$UNB_CHP'}' )" >> $statsFile
 	
@@ -208,7 +306,7 @@ Generate_UnboundStats () {
 	
 	"$SQLITE3_PATH" "$dbStats" < /tmp/unbound-stats.sql
 	
-	rm -f "$statsCHPFileJS"
+	[ -f $statsCHPFileJS ] && rm -f "$statsCHPFileJS"
 	WriteData_ToJS "/tmp/unbound-chp-daily.csv" "$statsCHPFileJS" "DatadivLineChartCacheHitPercentDaily"
 	WriteData_ToJS "/tmp/unbound-chp-weekly.csv" "$statsCHPFileJS" "DatadivLineChartCacheHitPercentWeekly"
 	WriteData_ToJS "/tmp/unbound-chp-monthly.csv" "$statsCHPFileJS" "DatadivLineChartCacheHitPercentMonthly"
@@ -224,6 +322,50 @@ Generate_UnboundStats () {
 	[ -f $statsAnswersFileJS ] && rm -f $statsAnswersFileJS
 	WriteUnboundStats_ToJS "barDataAnswers" $statsAnswersFileJS $raw_statsFile "num.answer.rcode.NOERROR" "num.answer.rcode.FORMERR" "num.answer.rcode.SERVFAIL" "num.answer.rcode.NXDOMAIN" "num.answer.rcode.NOTIMPL" "num.answer.rcode.REFUSED"
 	WriteUnboundLabels_ToJS "barLabelsAnswers" $statsAnswersFileJS "DNS Query completed successfully" "DNS Query Format Error" "Server failed to complete the DNS request" "Domain name does not exist  (including adblock if enabled)" "Function not implemented" "The server refused to answer for the query"
+
+	#generate data for top blocked domains
+	echo "Outputting top blocked domains..."
+	[ -f $statsTopBlockedFileJS ] && rm -f $statsTopBlockedFileJS
+	WriteUnboundSqlLog_ToFile "nx_domains" "domain" "count" "10" "/tmp/unbound-tbd.csv" "/tmp/unbound-tbd.sql"
+	"$SQLITE3_PATH" "$dbLogs" < /tmp/unbound-tbd.sql
+	WriteUnboundCSV_ToJS "/tmp/unbound-tbd.csv" "$statsTopBlockedFileJS" "barLabelsTopBlocked" "barDataTopBlocked"
+
+	#generate data for top 10 weekly replies from unbound
+	echo "Outputting top replies ..."
+	[ -f $statsTopRepliesFileJS ] && rm -f $statsTopRepliesFileJS
+	WriteUnboundSqlLog_ToFile "reply_domains" "domain, reply" "count" "10" "/tmp/unbound-topreplies.csv" "/tmp/unbound-topreplies.sql"
+	"$SQLITE3_PATH" "$dbLogs" < /tmp/unbound-topreplies.sql
+	WriteUnboundCSV_ToJS_2Labels "/tmp/unbound-topreplies.csv" "$statsTopRepliesFileJS" "barLabelsTopReplies" "barDataTopReplies"
+
+	#generate daily replies CSV
+	echo "Outputting daily replies ..."
+	[ -f $statsDailyRepliesFileJS ] && rm -f $statsDailyRepliesFileJS
+	whereString="WHERE date='""$(date '+%F')""'"
+	WriteUnboundSqlLog_ToFile "reply_domains" "domain, client_ip, reply" "count" "250" "/tmp/unbound-dailyreplies.csv" "/tmp/unbound-dailyreplies.sql" "$whereString"
+	"$SQLITE3_PATH" "$dbLogs" < /tmp/unbound-dailyreplies.sql
+	dos2unix "/tmp/unbound-dailyreplies.csv"
+	WriteUnboundCSV_ToJS_Table "/tmp/unbound-dailyreplies.csv" $statsDailyRepliesFileJS "LoadDailyRepliesTable" "DatadivTableDailyReplies"
+
+	#generate DNS Firewall Events (RPZ)
+	echo "Calculating DNS Firewall data..."
+	{
+		echo ".mode csv"
+		echo ".output /tmp/unbound-rpz-monthly.csv"
+		echo "select [date],SUM(count) from rpz_events GROUP BY date ORDER BY date;"
+	} > /tmp/unbound-rpz.sql
+	
+	"$SQLITE3_PATH" "$dbLogs" < /tmp/unbound-rpz.sql
+	[ -f $statsRPZFileJS ] && rm -f "$statsRPZFileJS"
+	WriteData_ToJS "/tmp/unbound-rpz-monthly.csv" "$statsRPZFileJS" "DatadivLineChartRPZHitsMonthly" "date-day"
+
+	#generate table data for all known RPZ hits
+	echo "Outputting DNS Firewall Hits ..."
+	[ -f $statsRPZHItsFileJS ] && rm -f $statsRPZHitsFileJS
+	whereString=""
+	WriteUnboundSqlLog_ToFile "rpz_events" "domain, client_ip, zone" "count" "250" "/tmp/unbound-rpzhits.csv" "/tmp/unbound-rpzhits.sql" "$whereString"
+	"$SQLITE3_PATH" "$dbLogs" < /tmp/unbound-rpzhits.sql
+	dos2unix "/tmp/unbound-rpzhits.csv"
+	WriteUnboundCSV_ToJS_Table "/tmp/unbound-rpzhits.csv" $statsRPZHitsFileJS "LoadRPZHitsTable" "DatadivTableRPZHits"
 
 	#cleanup temp files
 	rm -f "/tmp/unbound-"*".csv"
@@ -269,12 +411,22 @@ Auto_Cron(){
 			if [ "$STARTUPLINECOUNT" -eq 0 ]; then
 				cru a "$SCRIPT_NAME" "59 * * * * $SCRIPT_DIR/$SCRIPT_NAME_LOWER generate"
 			fi
+			STARTUPLINECOUNT=$(cru l | grep -c "$LOGSCRIPT_NAME")
+			
+			if [ "$STARTUPLINECOUNT" -eq 0 ]; then
+				cru a "$LOGSCRIPT_NAME" "57 * * * * $SCRIPT_DIR/$LOGSCRIPT_NAME_LOWER"
+			fi
 		;;
 		delete)
 			STARTUPLINECOUNT=$(cru l | grep -c "$SCRIPT_NAME")
 			
 			if [ "$STARTUPLINECOUNT" -gt 0 ]; then
 				cru d "$SCRIPT_NAME"
+			fi
+			STARTUPLINECOUNT=$(cru l | grep -c "$LOGSCRIPT_NAME")
+			
+			if [ "$STARTUPLINECOUNT" -gt 0 ]; then
+				cru d "$LOGSCRIPT_NAME"
 			fi
 		;;
 	esac
@@ -404,7 +556,8 @@ Unmount_WebUI(){
 	fi
 }
 
-ScriptHeader(){
+# $1 show commands
+ScriptHeader() { 
 	printf "\\n"
 	printf "##\\n"
 	printf "# ____ ___     ___.                            .___   _________ __          __          \\n"
@@ -416,10 +569,12 @@ ScriptHeader(){
 	printf "## by @juched - Generate Stats for GUI tab - %s                                         \\n" "$SCRIPT_VERSION"
 	printf "## with credit to @JackYaz for his shared scripts                                       \\n"
 	printf "\\n"
-	printf "unbound_stats.sh\\n"
-	printf "		install   - Installs the needed files to show UI and update stats\\n"
-	printf "		generate  - enerates statistics now for UI\\n"
-	printf "		uninstall - Removes files needed for UI and stops stats update\\n"
+	if [ ! -z $1 ]; then
+		printf "unbound_stats.sh\\n"
+		printf "		install   - Installs the needed files to show UI and update stats\\n"
+		printf "		generate  - enerates statistics now for UI\\n"
+		printf "		uninstall - Removes files needed for UI and stops stats update\\n"
+	fi
 }
 
 Download_File(){
@@ -480,10 +635,11 @@ Wait_For_Unbound() {
 
 #Main loop
 if [ -z "$1" ]; then
-	ScriptHeader
+	ScriptHeader show_commands
 	exit 0
 fi
 
+ScriptHeader
 case "$1" in
 	install)
 		Install_Dependancies
@@ -492,6 +648,7 @@ case "$1" in
 		Auto_Cron create
 		Mount_WebUI
 		Create_Dirs
+		sh /jffs/addons/unbound/unbound_log.sh
 		Generate_UnboundStats
 		exit 0
 	;;
@@ -520,6 +677,7 @@ case "$1" in
 		Unmount_WebUI
 		[ -f $installedMD5File ] && rm -f $installedMD5File
 		[ -f $dbStats ] &&  rm -f $dbStats
+		[ -f $dbLogs ] &&  rm -f $dbLogs
 		exit 0
 	;;
 esac
